move file to right path, use typedef everywhere
